cv-priv-app/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ann-Kathrin Kapfenstein", "Denis Moslavac", "Mario Ruoff"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.2.2"
rdflib = "^6.2.0"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
